gcp/video_processing/requirements.txt
functions-framework==3.*
opencv-python
google-cloud-storage
